Fix: Break loop after pickUp
This used to cause errors due to sometimes a value that was no longer in the
loop being called into pickUp. This also should eliminate false pick ups that
occured due to an item from powerSpots not having the same index value as its
position in allPowerUps. Both issues should now be fixed.