refactor: Encapsulate inherit in utils object
There is still a lot of encapsulation that this game could benefit
from. Ultimately these changes probably won't be made to this version
of the same as there are more fundamental improvements to be made to
the whole infastructure and its not entirely worth salvaging.

Built JS file is also updated here, those changes can of course be
ignored.